refactor: set sdk ID to com.drivequant.drivekit.flutter.tripanalysis
