+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis/example/android/app/build.gradle
@@ -21,14 +21,14 @@ android {
     }
 
 
-    namespace = "com.example.verygoodtrip.example"
+    namespace = "com.drivequant.drivekit.flutter.tripanalysis.example"
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
     }
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.example.verygoodtrip.example"
+        applicationId "com.drivequant.drivekit.flutter.tripanalysis.example"
         minSdkVersion 23
         targetSdkVersion 34
         versionCode 1
