+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis/example/android/app/src/debug/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.verygoodtrip.example">
+    package="com.drivequant.drivekit.flutter.tripanalysis.example">
     <!-- Flutter needs it to communicate with the running application
          to allow setting breakpoints, to provide hot reload, etc.
     -->
