+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis/example/android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" xmlns:tools="http://schemas.android.com/tools" package="com.example.verygoodtrip.example">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" xmlns:tools="http://schemas.android.com/tools" package="com.drivequant.drivekit.flutter.tripanalysis.example">
     <application
         android:label="example"
         android:name="${applicationName}"
