+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis/example/android/app/src/main/kotlin/com/drivequant/drivekit/flutter/tripanalysis/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.verygoodtrip.example
+package com.drivequant.drivekit.flutter.tripanalysis.example
 
 import io.flutter.embedding.android.FlutterActivity
