+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis/example/android/app/src/profile/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.verygoodtrip">
+    package="com.drivequant.drivekit.flutter.tripanalysis">
     <!-- Flutter needs it to communicate with the running application
          to allow setting breakpoints, to provide hot reload, etc.
     -->
