+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis_android/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.example.verygoodtrip">
+  package="com.drivequant.drivekit.flutter.tripanalysis">
 </manifest>
