+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis_android/android/src/main/kotlin/com/drivequant/drivekit/flutter/tripanalysis/DrivekitTripAnalysisPlugin.kt
@@ -1,4 +1,4 @@
-package com.drivekit.trip_analysis
+package com.drivequant.drivekit.flutter.tripanalysis
 
 import AndroidTripAnalysisApi
 import android.content.Context
